script exit 1 on error

+++ scripts/deploy_global_infra.sh
@@ -4,9 +4,23 @@ echo "deploying global insfrastructure"
 
 cd infrastructure/dev/global
 terraform init
-echo $?
+
+if [ "$?" -ne 0]; then
+   exit 1
+fi
+
 terraform plan -out terraformplan.out
+
+if [ "$?" -ne 0]; then
+   exit 1
+fi
+
 terraform apply terraformplan.out
+
+if [ "$?" -ne 0]; then
+   exit 1
+fi
+
 rm -rf terraformplan.out
 
 echo "Consumatum est..."
